# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: CivicLens AI API
  description: >
    REST API for the CivicLens AI civic-reporting platform — auth, progressive
    profiling, issue reporting, admin review, duplicate-flag resolution, and
    the trusted local-news feed. See /docs/API.md in the repo root for narrative
    walkthroughs of each flow.
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api
    description: Local development
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get the current user's profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
  /users/me/profile:
    patch:
      tags: [Users]
      summary: Progressive profiling — complete the profile after Google Sign-In
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username: { type: string, minLength: 3, maxLength: 30 }
                occupation: { type: string }
                city: { type: string }
                district: { type: string }
                photoUrl: { type: string, format: uri }
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        username: { type: string, nullable: true }
        occupation: { type: string, nullable: true }
        city: { type: string, nullable: true }
        district: { type: string, nullable: true }
        photo_url: { type: string, nullable: true }
        role: { type: string, enum: [citizen, admin, moderator] }
        profile_completed: { type: boolean }
...